Restructure operations
