+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/Cosine.kt
@@ -1,11 +1,11 @@
 package io.github.kapkekes.operations
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
-import io.github.kapkekes.cos
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
+import io.github.kapkekes.complex.cos
 
-object Cosine: Operation {
+class Cosine: Operation {
     override val token: String = "cos"
     override val arity: Int = 1
-    override val action: Action = { unit -> cos(unit.component1())}
+    override val action: Action = { unit -> cos(unit.component1()) }
 }
