+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/Logarithm.kt
@@ -1,10 +1,10 @@
 package io.github.kapkekes.operations
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
-import io.github.kapkekes.log
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
+import io.github.kapkekes.complex.log
 
-object Logarithm: Operation {
+class Logarithm: Operation {
     override val token: String = "log"
     override val arity: Int = 2
     override val action: Action = { pair -> log(pair.component2(), pair.component1()) }
