+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/Sine.kt
@@ -1,10 +1,10 @@
 package io.github.kapkekes.operations
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
-import io.github.kapkekes.sin
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
+import io.github.kapkekes.complex.sin
 
-object Sine: Operation {
+class Sine: Operation {
     override val token: String = "sin"
     override val arity: Int = 1
     override val action: Action = { unit -> sin(unit.component1()) }
