+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/SquareRoot.kt
@@ -1,10 +1,10 @@
 package io.github.kapkekes.operations
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
-import io.github.kapkekes.sqrt
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
+import io.github.kapkekes.complex.sqrt
 
-object SquareRoot: Operation {
+class SquareRoot: Operation {
     override val token: String = "sqrt"
     override val arity: Int = 1
     override val action: Action = { unit -> sqrt(unit.component1()) }
