+++ Task_1_5_1/src/main/kotlin/io/github/kapkekes/operations/Subtraction.kt
@@ -1,9 +1,9 @@
 package io.github.kapkekes.operations
 
-import io.github.kapkekes.Action
-import io.github.kapkekes.Operation
+import io.github.kapkekes.calculator.Action
+import io.github.kapkekes.calculator.Operation
 
-object Subtraction: Operation {
+class Subtraction: Operation {
     override val token: String = "-"
     override val arity: Int = 2
     override val action: Action = { pair -> pair.component1() - pair.component2() }
